+++ repository/src/main/java/jt/projects/repository/network/vk/VkNetworkRepository.kt
@@ -1,4 +1,4 @@
-package jt.projects.repository.network
+package jt.projects.repository.network.vk
 
 import jt.projects.model.*
